+++ src/nti/contentfragments/__init__.py
@@ -23,10 +23,7 @@ Other features:
 - Support for very configurable (optionally) event-based profanity
   censoring that integrates with nti.schema/zope.schema.
 
-.. $Id: __init__.py 58673 2015-02-02 17:52:58Z carlos.sanchez $
 """
 
-from __future__ import print_function, unicode_literals, absolute_import, division
+from __future__ import print_function, absolute_import, division
 __docformat__ = "restructuredtext en"
-
-logger = __import__('logging').getLogger(__name__)