+++ CHANGES.rst
@@ -2,11 +2,17 @@
  Changes
 =========
 
-1.1.2 (unreleased)
+1.2.0 (unreleased)
 ==================
 
-- Nothing changed yet.
+- Add support for Python 3.7. Note that ``datrie`` is not yet
+  available for Python 3.7.
 
+- Add support for PyPy3.
+
+- Add interfaces for all schema fields defined in
+  ``nti.contentfragments.schema`` and make the respective classes
+  implement them.
 
 1.1.1 (2018-06-29)
 ==================